+++ src/utils/opengles_display.c
@@ -586,12 +586,12 @@ void ogl_display_zoom(struct opengles_display* gldisp, float* params) {
 }
 
 #ifdef ANDROID
-JNIEXPORT void JNICALL Java_org_linphone_mediastream_video_display_OpenGLESDisplay_init(JNIEnv * env, jobject obj, jint ptr, jint width, jint height) {
+JNIEXPORT void JNICALL Java_org_linphone_mediastream_video_display_OpenGLESDisplay_init(JNIEnv * env, jobject obj, jlong ptr, jint width, jint height) {
 	struct opengles_display* d = (struct opengles_display*) ptr;
 	ogl_display_init(d, width, height);
 }
 
-JNIEXPORT void JNICALL Java_org_linphone_mediastream_video_display_OpenGLESDisplay_render(JNIEnv * env, jobject obj, jint ptr) {
+JNIEXPORT void JNICALL Java_org_linphone_mediastream_video_display_OpenGLESDisplay_render(JNIEnv * env, jobject obj, jlong ptr) {
 	struct opengles_display* d = (struct opengles_display*) ptr;
 	ogl_display_render(d, 0);
 }
